support object id by simpleflake

+++ attachments/models.py
@@ -31,7 +31,7 @@ class Attachment(models.Model):
     objects = AttachmentManager()
 
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
+    object_id = models.BigIntegerField()
     content_object = GenericForeignKey('content_type', 'object_id')
     creator = models.ForeignKey(settings.AUTH_USER_MODEL, related_name="created_attachments",
         verbose_name=_('creator'), on_delete=models.CASCADE)
